Rewind: Introduce concept of discarded events
@see #18944 - related preparatory work

Events in the activity log can be considered discarded from specific
perspectives when examining the list. This happens when an earlier
backup has been restored over the event. However, because we can restore
backups which hide other restore operations the determination can get a
little tricky.

In this PR we're introducing a new module to augment a stream of events
and update the `isDiscarded` value based on the inferred list of restore
operations and the perspective time, which at the moment is the "now."

Note:
 - There is some funny state going on with the `QueryManager` and I'd
   like to investigate why that is causing trouble. It may be obscuring
   the value of this property for older events or events on previous
   pages in the activity log view.
 - I've added dump styling to make discarded events jump out. If someone
   would like to help out then proper styling would be well appreciated,
   though in a separate PR would be best. The `is-discarded` CSS class
   will remain on the `activity-log-item` `<div>`.
 - We're exploring setting this value in the API itself. We're also
   exploring some interactive views such as recalculating when hovering
   over a possible restore point to give a better preview of what events
   would be discarded after restoring.

Todo:
 -[ ] Proper styling of discarded events
 -[ ] Figure out state glitches
 -[ ] Refactor state (can do in #18944 maybe)
 -[ ] Confirm this works on older pages of events
 -[ ] Fill in README