Data Layer: Allow Redux actions as network request handlers
Pretty often I see network requests made via `dispatchRequest()` and the
response handler for `onFailure` is nothing more than a notice to alert
our customers that something failed.

```js
const announceFailure = ( { dispatch } ) => errorNotice( translate(
'Sad' ) );
```

Well this is a lot of code. In many cases we just want to dispatch an
action. This is a simple thing and we should not require a function
where a simple object will do.

```js
dispatchRequest(
	fetchThing,
	storeThingInState,
	errorNotice( translate( 'Sad' ) ),
)
```

Actually there's one major flaw here: translations need to happen at the
time of dispatch, not at app boot as they would here, so that's a thing
to resolve.

Naively we could wrap `dispatchRequest()` but I think this is a terrible
idea.